% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\encoding{UTF-8}
\name{medreg}
\alias{medreg}
\title{Regeneration of Mediterranean vegetation data set}
\format{
An object of class \code{stratifiedvegdata} with 96 elements (plots), each of them consisting of a data.frame where rows correspond to species groups and columns correspond to vegetation strata. Abundance values are percentage cover.
}
\description{
A stratified vegetation data set containing with several plot records laid to assess vegetation recovery three years after a wildfire. Collected in 2012 by Miquel De Caceres and Albert Petit in Horta de Sant Joan (Catalonia, Spain).
}
\seealso{
\code{\link{CAP}}, \code{\link{plot.CAP}}, \code{\link{stratifyvegdata}}
}
\keyword{data}
