% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interclustdist.R
\name{interclustdist}
\alias{interclustdist}
\title{Distance between pairs of cluster centroids}
\usage{
interclustdist(x, cluster)
}
\arguments{
\item{x}{A site-by-site data matrix or an object of class \code{\link{dist}} containing the distance values between pairs of sites (plot records).}

\item{cluster}{A vector indicating the hard membership of each object in \code{x} to a set of groups. Can contain \code{NA} values.}
}
\value{
An object of class \code{\link{dist}} containing the distances between cluster centers.
}
\description{
Calculates the distance between pairs of cluster centroids, given a distance matrix and a cluster vector.
}
\author{
Miquel De \enc{Cáceres}{Caceres}, CREAF
}
