\name{medreg}
\alias{medreg}
\docType{data}
\title{Regeneration of Mediterranean vegetation data set}
\description{
  A stratified vegetation data set containing with several plot records laid to assess vegetation recovery three years after a wildfire. Collected in 2012 by Miquel De Caceres and Albert Petit in Horta de Sant Joan (Catalonia, Spain).
}
\usage{data(medreg)}
\format{
  An object of class 'stratifiedvegdata' with 96 elements (plots), each of them consisting of a data.frame where rows correspond to species groups and columns correspond to vegetation strata. Abundance values are percentage cover.
}
\seealso{\code{\link{CAP}}, \code{\link{plot.CAP}}}
\examples{
data(medreg)
}
\keyword{datasets}
