% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\encoding{UTF-8}
\name{treedata}
\alias{treedata}
\title{Synthetic vegetation data set with tree data}
\format{
A data frame where each row corresponds to a different tree. Columns are plot code, species identity, tree height, tree diameter and cover value.
}
\description{
A synthetic data set used to illustrate the stratification of data originally collected on an individual basis (e.g. forest inventory).
}
\seealso{
\code{\link{stratifyvegdata}}
}
\keyword{data}
