\name{treedata}
\alias{treedata}
\docType{data}
\title{Synthetic vegetation data set with tree data}
\description{
  A synthetic data set used to illustrate the stratification of data originally collected on an individual basis (e.g. forest inventory).
}
\usage{data(treedata)}
\format{
  A data frame where each row corresponds to a different tree. Columns are plot code, species identity, tree height, tree diameter and cover value.
}
\seealso{\code{\link{stratifyvegdata}}}
\keyword{datasets}
